+++ NEWS.md
@@ -23,6 +23,7 @@
 - Span borders more accurately show if the span's beginning or end are outside the visible area
 - Merged spans no longer stop at the screen boundaries
 - Lowered peak memory usage
+- "active" spans in goroutine timelines now have a label
 
 
 # v0.3.0 (2023-08-07)